added jwt tokens route guards cookies to save user token removed navbar from singin page

+++ Server/controller/businessLogic.ts
@@ -5,6 +5,7 @@ import {Request, Response} from "express";
 import {Credentials} from "../types/credentials";
 import {User} from "../types/user";
 import {credentialsSchema, userSchemas} from "../schemas/userSchemas";
+import {Auth} from "../services/auth/auth";
 
 
 export class BusinessLogic {
@@ -17,6 +18,15 @@ export class BusinessLogic {
     }
 
     static async getUsers(req: Request, res: Response): Promise<void> {
+        try {
+            const users: User[] = await DBAccess.selectMany('Users', {});
+            res.status(200).send(users);
+        } catch (e: any) {
+            res.status(500).send(e.message);
+        }
+    }
+    static async getUser(req: Request, res: Response): Promise<void> {
+        console.log(req.headers.token);
         try {
             const users: User[] = await DBAccess.selectMany('Users', {});
             res.status(200).send(users);
@@ -59,6 +69,7 @@ export class BusinessLogic {
                 password: credentials.password
             });
             if (user) {
+                user.token = Auth.createToken(user);
                 res.status(200).send(user);
             } else {
                 res.status(401).send('email or password is incorrect');
@@ -67,6 +78,7 @@ export class BusinessLogic {
             res.status(500).send(e.message);
         }
     }
+
     static async insertUser(req: Request, res: Response): Promise<void> {
         const user: User = req.body.user;
         const validation: Joi.ValidationResult = userSchemas.validate(user);
@@ -89,6 +101,11 @@ export class BusinessLogic {
 
     static async deleteUser(req: Request, res: Response): Promise<void> {
         const userId: string = req.body.data;
+        const token: string = req.body.token;
+        if (!Auth.isAdmin(token)) {
+            res.status(401).send('Unauthorized');
+            return
+        }
         try {
             await DBAccess.deleteOne('Users', {_id: new ObjectId(userId)});
         } catch (e) {
